Add files via upload
Added 2 files to run the scripts globally.
1_findMeritcoord_old.py
2_makeshape1_old.py

For a global run you cannot use a single 3sec ups or dir file.
Instead it is using the folder structure of merit
http://hydro.iis.u-tokyo.ac.jp/~yamadai/MERIT_Hydro/
(each directory 30 deg x 30 deg , each file 5deg x 5deg)
e.g
dir_n30w090
   n30w065_dir.tif
   n30w070_dir.tif

For getting the right point in the upstream map it uses the .tif map of the location
If the point is too close to the edge of a .tif it will expand the upstream area with the closest tiffs.

For the river network (dir). It takes an old shape file of the basin (could be also a raster) to detect the area to use for calculating the fine resolution basin shapefile. Or if you do not have a shape or raster, it uses an rough estimate of the area